Added dynamic password requirement check and improved register page layout
added a function in a new file that checks the password being entered in the registration form meets the requirements which is called in register.ejs. changed a mistake of mislabelled type in the name section of form to make styling accurate. added new css classes to make the registration form look good.