added usertypes for instructors and customers
added usertypes for instructors and customers to make sure that only instructors can add events, next steps will be to allow customers to sign up for events and instructors able to delete events. i also want customers to be able to request private lessons but not sure how i would do that.